Move Box to Shapes namespace
