+++ osu.Framework.Testing/TestCase.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using osu.Framework.Extensions.TypeExtensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Testing.Drawables.StepButtons;
 using osu.Framework.Threading;
 using OpenTK;
