+++ osu.Framework.VisualTests/Tests/TestCaseFillModes.cs
@@ -7,6 +7,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Input;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Testing;
