+++ osu.Framework.VisualTests/Tests/TestCaseNestedHover.cs
@@ -3,10 +3,10 @@
 
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Input;
 using OpenTK;
 using OpenTK.Graphics;
-using osu.Framework.Graphics.Sprites;
 using osu.Framework.Testing;
 
 namespace osu.Framework.VisualTests.Tests
