+++ osu.Framework.VisualTests/Tests/TestCaseScreen.cs
@@ -4,6 +4,7 @@
 using System;
 using osu.Framework.Screens;
 using osu.Framework.Graphics;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.MathUtils;
