+++ osu.Framework.VisualTests/Tests/TestCaseSmoothedEdges.cs
@@ -3,6 +3,7 @@
 
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Testing;
 using OpenTK;
 using OpenTK.Graphics;
