+++ osu.Framework.VisualTests/Tests/TestCaseTabControl.cs
@@ -9,6 +9,7 @@ using OpenTK.Graphics;
 using osu.Framework.Extensions;
 using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Framework.Graphics;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Testing;
