+++ osu.Framework/Graphics/Containers/ScrollContainer.cs
@@ -7,7 +7,7 @@ using osu.Framework.Input;
 using osu.Framework.MathUtils;
 using OpenTK;
 using OpenTK.Graphics;
-using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Shapes;
 
 namespace osu.Framework.Graphics.Containers
 {
