+++ osu.Framework/Graphics/Cursor/CursorContainer.cs
@@ -2,9 +2,9 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Input;
 using OpenTK;
-using osu.Framework.Graphics.Sprites;
 using OpenTK.Graphics;
 
 namespace osu.Framework.Graphics.Cursor
