+++ osu.Framework/Graphics/Performance/FpsDisplay.cs
@@ -3,6 +3,7 @@
 
 using OpenTK.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.MathUtils;
 using osu.Framework.Timing;
