+++ osu.Framework/Graphics/Performance/FrameStatisticsDisplay.cs
@@ -7,6 +7,7 @@ using System.Drawing;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.OpenGL.Textures;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Input;
