+++ osu.Framework/Graphics/Shapes/Box.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
+using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 
-namespace osu.Framework.Graphics.Sprites
+namespace osu.Framework.Graphics.Shapes
 {
     /// <summary>
     /// A simple rectangular box. Can be colored using the <see cref="Drawable.Colour"/> property.
