+++ osu.Framework/Graphics/UserInterface/DropdownHeader.cs
@@ -3,7 +3,7 @@
 
 using OpenTK.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Shapes;
 using osu.Framework.Input;
 
 namespace osu.Framework.Graphics.UserInterface
